backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-dotenv
feedparser
beautifulsoup4
nltk
vaderSentiment
empath
openai
pymongo
newspaper3k
lxml_html_clean
httpx
apscheduler


######

fastapi
uvicorn
transformers
torch
transformers
requests
newspaper3k
openai
pandas
python-multipart
googlesearch-python==1.1
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-cloud-storage
google-cloud-bigquery
google-cloud-bigquery-storage


# ==== FastAPI Backend ====
fastapi
uvicorn
pandas
openpyxl
python-dotenv
feedparser
beautifulsoup4

# ==== NLP & Sentiment ====
nltk
vaderSentiment
empath
openai

# ==== Optional for development ====
requests


fastapi
uvicorn
pandas
openpyxl
python-dotenv
feedparser
beautifulsoup4
nltk
vaderSentiment
empath
openai
pymongo
